readme

+++ contracts/Vether.sol
@@ -3,10 +3,10 @@ pragma solidity 0.6.4;
 interface ERC20 {
     function totalSupply() external view returns (uint256);
     function balanceOf(address account) external view returns (uint256);
-    function transfer(address, uint) external returns (bool);
+    function transfer(address, uint256) external returns (bool);
     function allowance(address owner, address spender) external view returns (uint256);
-    function approve(address, uint) external returns (bool);
-    function transferFrom(address, address, uint) external returns (bool);
+    function approve(address, uint256) external returns (bool);
+    function transferFrom(address, address, uint256) external returns (bool);
     event Transfer(address indexed from, address indexed to, uint256 value);
     event Approval(address indexed owner, address indexed spender, uint256 value);
     }
@@ -197,7 +197,7 @@ contract Vether is ERC20 {
     // ################-REMOVE_THIS_FOR_MAINNET-##########################
     //======================================WITHDRAWAL======================================//
     // Used to efficiently track participation in each era
-    function getDaysContributedForEra(address member, uint256 era) public view returns(uint){
+    function getDaysContributedForEra(address member, uint256 era) public view returns(uint256){
         return mapMemberEra_Days[member][era].length;
     }
     // Call to withdraw a claim
